* commit '6f17adbabc329db1671c2fa6fd9e25cbf6a3fa8e':
  AArch64: Use long for pointers in media classes

+++ media/java/android/media/AudioTrack.java
@@ -221,13 +221,13 @@ public class AudioTrack
      * Accessed by native methods: provides access to C++ AudioTrack object.
      */
     @SuppressWarnings("unused")
-    private int mNativeTrackInJavaObj;
+    private long mNativeTrackInJavaObj;
     /**
      * Accessed by native methods: provides access to the JNI data (i.e. resources used by
      * the native AudioTrack object, but not stored in it).
      */
     @SuppressWarnings("unused")
-    private int mJniData;
+    private long mJniData;
 
 
     //--------------------------------------------------------------------------
